cleaning of the code

+++ src/main/java/com/game/spaceinvaders/GameObject.java
@@ -4,53 +4,43 @@ package com.game.spaceinvaders;
 import javafx.beans.property.*;
 
 public class GameObject {
-
-    //TODO: možná udělat "Wall" tridu jako podtridu GameObject (oproti tomu ma navic jen lives
-
     private final BooleanProperty active;
-    private final DoubleProperty row;
-    private final DoubleProperty col;
+    private final IntegerProperty row;
+    private final IntegerProperty col;
 
-
-    //private final IntegerProperty size;
     private final IntegerProperty width;
     private final IntegerProperty height;
 
-
-    //TODO: kouknout proc je nekde "int" a v this je  SimpleDoubleProperty
-    //trida pro monstra i hrace
     public GameObject(int row, int col, int width, int height) {
-        this.row = new SimpleDoubleProperty(row);
-        this.col = new SimpleDoubleProperty(col);
+        this.row = new SimpleIntegerProperty(row);
+        this.col = new SimpleIntegerProperty(col);
         //this.size = new SimpleIntegerProperty(size);
         this.active = new SimpleBooleanProperty(true);
 
         this.width = new SimpleIntegerProperty(width);
         this.height = new SimpleIntegerProperty(height);
     }
-    //x
-    public final DoubleProperty rowProperty() {
+
+    public final IntegerProperty rowProperty() {
         return this.row;
     }
-    public final double getRow() {
+    public final int getRow() {
         return this.rowProperty().get();
     }
-    public final void setRow(final double row) {
+    public final void setRow(final int row) {
         this.rowProperty().set(row);
     }
 
-    //y
-    public final DoubleProperty colProperty() {
+    public final IntegerProperty colProperty() {
         return this.col;
     }
-    public final double getCol() {
+    public final int getCol() {
         return this.colProperty().get();
     }
-    public final void setCol(final double col) {
+    public final void setCol(final int col) {
         this.colProperty().set(col);
     }
 
-    //active
     public final BooleanProperty activeProperty() {
         return this.active;
     }
@@ -61,7 +51,6 @@ public class GameObject {
         this.activeProperty().set(active);
     }
 
-    //height
     public final IntegerProperty heightProperty() {
         return this.height;
     }
@@ -72,7 +61,6 @@ public class GameObject {
         this.heightProperty().set(height);
     }
 
-    //width
     public final IntegerProperty widthProperty() {
         return this.width;
     }
